Some moving
Moved LocationDAO in the DAO package
