+++ src/main/java/edu/wpi/teamg/DAOs/LocationDAO.java
@@ -1,4 +1,4 @@
-package edu.wpi.teamg;
+package edu.wpi.teamg.DAOs;
 
 import java.sql.SQLException;
 import java.util.List;
