+++ src/main/java/edu/wpi/teamg/EdgeDAO.java
@@ -1,3 +1,5 @@
 package edu.wpi.teamg;
 
+import edu.wpi.teamg.DAOs.LocationDAO;
+
 public interface EdgeDAO extends LocationDAO {}
