+++ src/main/java/edu/wpi/teamg/LocationNameDAO.java
@@ -1,3 +1,5 @@
 package edu.wpi.teamg;
 
+import edu.wpi.teamg.DAOs.LocationDAO;
+
 public interface LocationNameDAO extends LocationDAO {}
